+++ src/skins/DefaultSkin.js
@@ -204,14 +204,6 @@ define([
     }
 
     /**
-     * Gets the {@link ActiveBox} used to display the main messages of activities
-     * @returns {ActiveBox}
-     */
-    getMsgBox() {
-      return this.msgBox
-    }
-
-    /**
      * Enables or disables the `tabindex` attribute of the main buttons. Useful when a modal dialog
      * overlay is active, to avoid direct access to controls not related with the dialog.
      * @param {boolean} status - `true` to make main controls navigable, `false` otherwise